feat: ensure data cannot be reset in parallel

+++ front-end/src/main/services/localUser/index.ts
@@ -22,10 +22,19 @@ import { setPrismaClient, createPrismaClient } from '@main/db/prisma';
 export const userStorageFolderName = 'User Storage';
 export const getUserStorageFolderPath = (email: string) => `User Storage/${email}`;
 
+let resettingDatabase = false;
+
 export const resetDatabase = async () => {
-  await deleteDatabase();
-  setPrismaClient(createPrismaClient());
-  await initDatabase();
+  if (resettingDatabase) return;
+
+  try {
+    resettingDatabase = true;
+    await deleteDatabase();
+    setPrismaClient(createPrismaClient());
+    await initDatabase();
+  } finally {
+    resettingDatabase = false;
+  }
 };
 
 export const resetData = async () => {
